+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix content item categories [#67](https://github.com/rokwire/content-building-block/issues/67)
 
 ## [1.2.0] - 2022-06-10
 ### Changed